refactor(term): split oversized files and group the runtime input pipeline
Bring funput-term to the repo standard (<=150 LOC/file, <=5 files/dir) with a
pure internal reorganization — no public API or behavior change.

- runtime/input.rs (262 LOC) -> runtime/driver/classify.rs: extract the inline
  classifier tests to runtime/driver/classify/tests.rs (matching the existing
  driver/tests/ pattern)
- config/mod.rs (176 LOC) -> extract the inline tests to config/tests.rs
- Group the coupled input pipeline under runtime/driver/: the loop
  (driver.rs -> driver/mod.rs), the byte classifier (input.rs -> driver/classify.rs,
  renamed for clarity), and inject (inject.rs -> driver/inject.rs). This takes
  runtime/ from 7 files down to 4; driver/ has 4.

Also fixes pre-existing rustdoc private-intra-doc-link warnings in lib.rs and
config/mod.rs so the crate passes `cargo doc -D warnings`, and updates both
READMEs' module maps.

runtime's driver/input/inject were already crate-internal, and the public API
(runtime::{run, Options, Status}, config::*, install::*) is unchanged, so
funput-cli is untouched. Git tracks the moves as renames.
